Ya sirve la base en la nube

+++ database/migrations/2016_03_04_234203_create_platillos_table.php
@@ -16,7 +16,7 @@ class CreatePlatillosTable extends Migration {
 			$table->increments('id');
           	$table->string('nombre',150);
           	$table->string('descripcion',350);
-          	$table->double('precio');
+          	$table->float('precio');
           	$table->timestamps();
         });
